+++ src/main.jsx
@@ -3,15 +3,4 @@ import ReactDOM from "react-dom/client";
 import App from "./App";
 import "./index.css";
 
-import { Provider } from "react-redux";
-import store from "./redux/store";
-
-import { BrowserRouter } from "react-router-dom";
-
-ReactDOM.createRoot(document.getElementById("root")).render(
-  <Provider store={store}>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </Provider>
-);
+ReactDOM.createRoot(document.getElementById("root")).render(<App />);
